wip: adding `options` parameters to `keys` and `nodes` domain classes
[ci skip]
