+++ src/keys/types.ts
@@ -262,6 +262,35 @@ type CertManagerChangeData = {
   recoveryCode?: RecoveryCode;
 };
 
+/**
+ * Used by the PolykeyAgent for it's top level options
+ */
+type KeysOptions = KeyRingOptions & CertManagerOptions
+
+/**
+ * Options for the KeyRing
+ */
+type KeyRingOptions = {
+  passwordOpsLimit?: PasswordOpsLimit;
+  passwordMemLimit?: PasswordMemLimit;
+  strictMemoryLock: boolean;
+} & (
+  {} | {
+  recoveryCode: RecoveryCode;
+} | {
+  privateKey: PrivateKey;
+} | {
+  privateKeyPath: string;
+});
+
+/**
+ * Options for the CertManager
+ */
+type CertManagerOptions = {
+  certDuration: number;
+  certRenewLeadTime: number;
+};
+
 export type {
   BufferLocked,
   Key,
@@ -298,6 +327,9 @@ export type {
   CertificatePEM,
   CertificatePEMChain,
   CertManagerChangeData,
+  KeysOptions,
+  KeyRingOptions,
+  CertManagerOptions,
 };
 
 export type { CertId, CertIdString, CertIdEncoded } from '../ids/types';
